backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.10.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0
aiomysql>=0.2.0
oss2>=2.18.0
httpx>=0.24.0
pytest>=7.0.0
greenlet>=3.0.0
email-validator>=2.0.0
pypdfium2
openai
Pillow
